logger implemented on AudioHandler.java, Test edited and separeted from each other.
